requirements.txt
python-telegram-bot>=20.7
aiohttp==3.9.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
schedule==1.2.1
selenium==4.16.0
webdriver-manager==4.0.1
undetected-chromedriver==3.5.4
